Transferred from the Adobe Open Source Subversion Repository
Revision: 20217
Author:   klin@adobe.com
Date:     2011-02-09T00:00:08.304340Z

Log Message: 
-----------
Fix dispatching grid mouseUp and click events from Grid. Click was sometimes not being dispatched because FP checks the target underneath the mouse. In some cases between the edge of the renderer and separator, the target would change between the background and the selection indicator causing a mouse click from the player not to be dispatched. Fix is to check for a mouse up in the same cell. Also, in some cases when dragging in and out of the grid, mouseUp would be dispatched twice. The roll out handler was setting the dragInProgress flag to false which falsely indicated that the drag had completed when in fact the drag was still in progress.

QE notes: No
Doc notes: No
Bugs: SDK-29385, SDK-29409
Reviewer: Hans, Carol
Tests run: cyclone server
Is noteworthy for integration: No

Modified Paths:
--------------
/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/Grid.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1279032 13f79535-47bb-0310-9956-ffa450edef68