Transferred from the Adobe Open Source Subversion Repository
Revision: 12964
Author:   jimurphy@adobe.com
Date:     2009-12-15T18:41:33.901783Z

Log Message: 
-----------
Partial revert of rev# 12926.
frameworks/projects/spark/src/spark/components/Group.as

A Group cannot share its display object when layering is in effect or blendMode is not normal. Blend mode is not normal when it is explicitly set OR blendMode is auto and a fractional alpha is at play. 

Bugs: http://bugs.adobe.com/jira/browse/SDK-24667

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/spark/src/spark/components/Group.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1279220 13f79535-47bb-0310-9956-ffa450edef68