Transferred from the Adobe Open Source Subversion Repository
Revision: 4550
Author:   dsubrama@adobe.com
Date:     2009-01-16T01:26:55.828631Z

Log Message: 
-----------
We are now going to allow proxied properties be specified in the component skin. Ryan is to write this up as part of his proxied property communication document.

This covers the specific case where Thermo would like to have the default itemRenderer used by an FxList control settable in the skin, instead of setting it on the component instance. 

Checkintests: Pass
Reviewer: Ryan
Mustella tests: Pass

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/mx/components/FxList.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/mx/skins/spark/FxListSkin.mxml</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1279553 13f79535-47bb-0310-9956-ffa450edef68