Transferred from the Adobe Open Source Subversion Repository
Revision: 12300
Author:   jszeto@adobe.com
Date:     2009-11-30T22:51:26.716735Z

Log Message: 
-----------
PopUpAnchor effects bug fix

http://bugs.adobe.com/jira/browse/SDK-24187 - PopUpAnchor removal changes width/height of popup, causing artifacts with any effects

Changed PopUpAnchor logic so that we only capture the popUp size the first time it is popped up. We no longer restore the value when the popup is removed. Instead, we restore the value when we open the popUp and we aren't forcing the size to match the anchor. 

QE notes: None
Doc notes: None
Bugs: SDK-24187
Reviewer: Chet
Tests run: PopUpAnchor, ComboBox, DropDownList
Is noteworthy for integration: No

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/spark/src/spark/components/PopUpAnchor.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1279877 13f79535-47bb-0310-9956-ffa450edef68