Transferred from the Adobe Open Source Subversion Repository
Revision: 6626
Author:   gosmith@adobe.com
Date:     2009-05-05T01:45:01.568612Z

Log Message: 
-----------
Integrated a private build of TLF from Rich.

This build has an additional tabbing-related fix on top of TLF Build 424. We decided to take 424 because it allows all event handling in TextContainerManager to be overridden, which may be important for fixing beta bugs. But Rich also fixed a new bug that I found today: if you had moused over a TextInput or TextArea, then when you tabbed to it you couldn't enter text. This build passed the TextInput and TextArea Mustella tests. We'll take the official 425 tomorrow which will be 424 plus this new fix.

QE Notes: None
Doc Notes: None
Bugs: None
Reviewer: Carol

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/in/tlf/textLayout.swz</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/spark/primitives/RichEditableText.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/libs/textLayout.swc</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/spark/primitives/supportClasses/RichEditableTextContainerManager.as</path>
<path kind="file" action="M">/flex/sdk/trunk/in/tlf/textLayout.swf</path>
<path kind="file" action="M">/flex/sdk/trunk/in/tlf/tlf.properties</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1280092 13f79535-47bb-0310-9956-ffa450edef68