Transferred from the Adobe Open Source Subversion Repository
Revision: 6695
Author:   gosmith@adobe.com
Date:     2009-05-06T20:50:19.055934Z

Log Message: 
-----------
Fixed code for drawing debugging backgrounds in SimpleText, RichText, and RichEditableText.

In SimpleText and RichText, it shouldn't do graphics.clear(), and should draw at (drawX, drawY), in order to work properly when TextGraphicElements are sharing a single DisplayObject.

In RichEditableText I moved the static var that controls background-drawing from the subclass of TextContainerManager that actually does the drawing to the RichEditableText class itself, to be consistent with how you set it on SimpleText and RichText.

QE Notes: None
Doc Notes: None
Bugs: None
Reviewer: Ryan

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/spark/primitives/RichEditableText.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/spark/primitives/RichText.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/spark/primitives/supportClasses/RichEditableTextContainerManager.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/spark/primitives/SimpleText.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1280096 13f79535-47bb-0310-9956-ffa450edef68