Transferred from the Adobe Open Source Subversion Repository
Revision: 8576
Author:   cframpto@adobe.com
Date:     2009-07-15T12:19:18.531995Z

Log Message: 
-----------
RichEditableText will dispatch all the FlexEvent.COMMIT_VALUE events and SkinnableTextBase will redispatch them for TextInput and TextArea.  Dispatched on focusOut, set text/content/textFlow, and for the insertText() and appendText() API calls.

Also changed a few validateNow() calls to validateProperties() because that is all that is needed and it is a much safer call.

QA Notes:
Doc Notes: SDK-21991
Reviewers: Alex

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/spark/src/spark/primitives/RichEditableText.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/spark/src/spark/components/TextArea.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/spark/src/spark/components/supportClasses/SkinnableTextBase.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/spark/src/spark/components/TextInput.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1280150 13f79535-47bb-0310-9956-ffa450edef68