Transferred from the Adobe Open Source Subversion Repository
Revision: 9407
Author:   cframpto@adobe.com
Date:     2009-08-19T22:11:34.142583Z

Log Message: 
-----------
Reordering the if-statement in isMeasureFixed() exposed that hostFormat can be null when measure() is called if styles changed and measure is done before the next commitProperties.  This states test exposed this.

We should rethink if we want to clear hostFormat rather than have a hostFormatChanged flag.  If there is no hostFormat at measure then it has to be fixed because there is no line break format to check for auto-size.

QE notes: 
Doc notes: 
Bugs: SDK-22779
Reviewer: Gordon
Tests run: checkintests
Is noteworthy for integration: no

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/spark/src/spark/primitives/RichEditableText.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1280170 13f79535-47bb-0310-9956-ffa450edef68