Transferred from the Adobe Open Source Subversion Repository
Revision: 9592
Author:   cframpto@adobe.com
Date:     2009-08-25T21:06:23.898916Z

Log Message: 
-----------
Fix SDK-22469 again (even though it isn't open) which I realized I rebroke when I removed the code to set contentWidth/Height to 0 in measure().  We don't want the scroller using old content values if we are remeasuring.

Removed code added last week to compositionCompleteHandler which if scrolling and toFit, reported the contentWidth as the compositionWidth in an attempt to make sure the scroller didn't add horizontal scroll bars.  This was the wrong approach.

QE notes: we really need a test in mustella for SDK-22469
Doc notes: 
Bugs: SDK-22469
Reviewer: Gordon and Hans
Tests run: TextArea
Is noteworthy for integration: no

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/spark/src/spark/primitives/RichEditableText.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1280173 13f79535-47bb-0310-9956-ffa450edef68