Transferred from the Adobe Open Source Subversion Repository
Revision: 11943
Author:   cframpto@adobe.com
Date:     2009-11-18T17:39:35.268598Z

Log Message: 
-----------
When remeasuring text for auto-size, set the compositionHeight correctly in the case were the height is constrained but the width is not.

Due to TLF bug, an image embedded in a compiled text flow does not finish loading until InlineGraphicElement.mustUseComposer() is called.  In our textContainerManager_inlineGraphicStatusChangeHandler, when we get the SIZE_PENDING event force the state transition.  Watson bug #2486102 logged against Vellum V2.

Change validateNow() in api calls to validateProperties() because that is really all we need and it reduces the risk of an infinite loop regarding measure/damage/measure/damage/etc.

QE notes: 
Doc notes: 
Bugs: SDK23973, SDK-24288, SDK-23638
Reviewer: Gordon
Tests run: checkintests
Is noteworthy for integration: no

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/spark/src/spark/components/RichEditableText.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1280195 13f79535-47bb-0310-9956-ffa450edef68