Transferred from the Adobe Open Source Subversion Repository
Revision: 12362
Author:   cframpto@adobe.com
Date:     2009-12-02T13:30:12.251693Z

Log Message: 
-----------
Revert the part of svn 12322 that added support for custom context menus to TextArea, TextInput and RichEditableText.  This needs to be reworked so that it builds and works correctly for AIR apps as well.  

Apparently you can't override the contextMenu setter/getter because they use a ContextMenu on Flash and a NativeMenu on AIR so the signatures will only match one environment depending on which menu class you use.

Reopened SDK-23926.

QE notes:
Doc notes: 
Bugs: SDK-23926
Reviewer:
Tests run: checkintests
Is noteworthy for integration: no

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/spark/src/spark/components/supportClasses/SkinnableTextBase.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/spark/src/spark/components/RichEditableText.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1280201 13f79535-47bb-0310-9956-ffa450edef68