Transferred from the Adobe Open Source Subversion Repository
Revision: 17794
Author:   cframpto@adobe.com
Date:     2010-09-21T21:49:15.794044Z

Log Message: 
-----------
When autoSizing, the code in measure resets TLF's vsp and hsp to 0, since it assumes the text will wrap and be visible.  If autoSize with lineBreak="explicit", as Darrell wants to do for the DataGrid cell editor, the text may need to scroll so the scroll positions should not be reset.

QE notes: 
Doc notes: 
Bugs: 
Reviewer: Darrell
Tests run: RichEditableText (the only tests that test autoSize), checkintests
Is noteworthy for integration:

Modified Paths:
--------------
/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/RichEditableText.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1280240 13f79535-47bb-0310-9956-ffa450edef68