Transferred from the Adobe Open Source Subversion Repository
Revision: 18147
Author:   cframpto@adobe.com
Date:     2010-10-14T12:22:33.142015Z

Log Message: 
-----------
When displayAsPassword is changed, TLF's text is reset which reinitializes TLF's textFlow.  Make sure to reinitialize our version of the textFlow and the textFlow generation numbers so that changes to the textFlow will not be dropped in the damageHandler.

QE notes: 
Doc notes: 
Bugs: SDK-28157
Reviewer: Alex
Tests run: cyclone
Is noteworthy for integration:

Modified Paths:
--------------
/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/RichEditableText.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1280242 13f79535-47bb-0310-9956-ffa450edef68