Transferred from the Adobe Open Source Subversion Repository
Revision: 18289
Author:   cframpto@adobe.com
Date:     2010-10-22T18:37:30.454222Z

Log Message: 
-----------
Partial implementation of preserveSelection.  If preserveSelecton is true, the selection is only preserved if selectionMode is "singleCell" or "singleRow".  We are still trying to figure out if we can handle "multipleCells" and "multipleRows" efficiently enough to allow a large number of selected items to be preserved.

Fix error in RichEditableText ASdoc.

QE notes: right now, only preserves selection if selectionMode="singleCell" or "singleRow"
Doc notes: 
Bugs: 
Reviewer: Hans
Tests run: cyclone - Jody excluded the 6 tests that fail because of no "mutipleCell" or "multipleRow" support for preserveSelection
Is noteworthy for integration:

Modified Paths:
--------------
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/RichEditableText.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/gridClasses/GridSelection.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1280247 13f79535-47bb-0310-9956-ffa450edef68