Transferred from the Adobe Open Source Subversion Repository
Revision: 18855
Author:   cframpto@adobe.com
Date:     2010-11-30T15:26:56.538530Z

Log Message: 
-----------
Fix another RichEditableText/Scroller looping bug by tweaking the existing code in adjustContentBoundsForScroller() which attempts to prevent looping.

For a given composition width, if there is not enough height for an inline graphic it does not seem to be included in the content measurement, and the contentWidth and contentHeight can decrease dramatically.  This can happen if there are auto scrollbars coming and going and causing reflow of the text to account for the space taken up or given back by the presence or absence of scrollbars.

QE notes: 
Doc notes: 
Bugs: SDK-28592 
Reviewer:  Hans
Tests run: cyclone
Is noteworthy for integration:

Modified Paths:
--------------
/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/RichEditableText.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1280251 13f79535-47bb-0310-9956-ffa450edef68