Transferred from the Adobe Open Source Subversion Repository
Revision: 3791
Author:   jszeto@adobe.com
Date:     2008-10-21T20:23:09.861725Z

Log Message: 
-----------
Fix bug when resetting stylesheet on FxApplication

SDK-17691 RTE loading styles

When the FxApplication replaces its style sheet, it unloads the previous skin and loads in a new one. With the new skin, it creates a new instance of the contentGroup. However, when the content is added to the contentGroup, it has already been parented by the previous version of the contentGroup. This causes the RTE because the content is not in the correct state. 

The solution is to set the old contentGroup's content property to null when we remove the skin. We also force the contentGroup to validate immediately because the new contentGroup was processing its content array before the oldGroup had removed the content as children. 

QE Notes: Needs testing
Doc Notes: none
Bugs: SDK-17691
Reviewer: Glenn

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/flex4/src/mx/components/FxContainer.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1280581 13f79535-47bb-0310-9956-ffa450edef68