Transferred from the Adobe Open Source Subversion Repository
Revision: 4035
Author:   jszeto@adobe.com
Date:     2008-11-06T23:38:55.516366Z

Log Message: 
-----------
RadialGradient and GradientEntry bug fixes

SDK-17745 - MXML application with multiple identical gradient fills doesn't render correctly

RadialGradient now takes into account the x/y position of the graphic element's bounds.

SDK-17652 - AI import; non-default angle and location of gradients are not retained on import of .ai or .fxg.

RadialGradient now rotates as expected. Previously, we set rotation using the Matrix.createGradientBox function. But this only rotate the gradients themselves and not the shape of the gradient as defined by scaleX and scaleY. Instead, we need to translate the center of the gradient to 0,0, rotate it, and then translate it back to the original location.  

SDK-17654 - GradientEntry ratio behavior is inconsistent between MXML and ActionScript

Changed the default value for the ratio parameter in the GradientEntry constructor from -1 to NaN. GradientEntry was probably written before we had support for NaN in the language. 


Added a couple of comments and removed some dead code.  

QE Notes: Please test rotation on RadialGradient and graphic elements with top/left or x/y defined and a RadialGradient fill. 
Doc Notes: Change default value of ratio parameter for GradientEntry constructor. 
Bugs: SDK-17745, SDK-17652, SDK-17654
Reviewer:

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/framework/src/mx/graphics/GradientEntry.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/mx/components/Group.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/mx/components/baseClasses/GroupBase.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/mx/components/FxContainer.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/mx/graphics/graphicsClasses/GraphicElement.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/framework/src/mx/graphics/RadialGradient.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1280585 13f79535-47bb-0310-9956-ffa450edef68