Transferred from the Adobe Open Source Subversion Repository
Revision: 9292
Author:   rfrishbe@adobe.com
Date:     2009-08-13T23:44:55.824541Z

Log Message: 
-----------
In a SkinnableContainer, if no content has been explicitly specified (via MXML or addElement and the other mutation APIs), then the SkinnableContainer, doesn't own the content, and the default specified in the skin will be left as-is.  This already happens for other proxied properties, so for the content, we now do it as well.

QE notes: -
Doc notes: -
Bugs: SDK-22688 
Reviewer: Glenn
Tests run: checkintests, mustella SkinnableComponent, SkinnableContainer
Is noteworthy for integration: No

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/spark/src/spark/components/SkinnableContainer.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1280626 13f79535-47bb-0310-9956-ffa450edef68