Transferred from the Adobe Open Source Subversion Repository
Revision: 11316
Author:   rfrishbe@adobe.com
Date:     2009-10-30T18:27:07.159086Z

Log Message: 
-----------
Fixing VideoPlayer bug: when VideoPlayer is inside a Vbox and is returning from full-screen mode, it RTEs.  At this point, the VideoPlayer was a direct descendent of the Application, but now it’s getting added back to it’s parent.  If that parent is a Halo container and the Application is a Spark Application, it will RTE because the Halo container will try to call removeChild() on the Spark Application.  Instead, I just handle the removing of the component in the VideoPlayer code and call removeElement() when appropriate.

QE notes: -
Doc notes: -
Bugs: SDK-23857
Reviewer: Glenn
Tests run: checkintests, mustella VideoPlayer
Is noteworthy for integration: No

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/spark/src/spark/components/VideoPlayer.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1281190 13f79535-47bb-0310-9956-ffa450edef68