Transferred from the Adobe Open Source Subversion Repository
Revision: 16696
Author:   cframpto@adobe.com
Date:     2010-06-28T14:48:27.485336Z

Log Message: 
-----------
The following changes apply to Spark DataGrid only.   Code reviews have been deferred until the implementation is closer to completion. 
 
Initial pass at getting the column header bar mouse event handlers framework set up for both header and separator events.  The header events use the grid events and there is a new set of events for the separators (in GridEvent for now).  In Grid, added resizeableColumns.  For a column to resize, this must be true and the individual column must be resizeable.  Added getColumnIndexAt() to GridLayout.  Added columnResizeIndicator to the CHB skin.

Modified Paths:
--------------
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/Grid.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/events/GridEvent.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/ColumnHeaderBar.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/skins/spark/ColumnHeaderBarSkin.mxml</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/supportClasses/GridLayout.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1281604 13f79535-47bb-0310-9956-ffa450edef68