Transferred from the Adobe Open Source Subversion Repository
Revision: 16754
Author:   cframpto@adobe.com
Date:     2010-07-01T01:10:56.986435Z

Log Message: 
-----------
The following changes apply to Spark DataGrid only.   Code reviews have been deferred until the implementation is closer to completion. 
 
In GridLayout.initializeItemRenderers, if the height for a row changes, make a 2nd pass thru the item renders in that row to resize them to the new height.  Use the new height when determining which rows are visible.

Modified Paths:
--------------
/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/supportClasses/GridLayout.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1281610 13f79535-47bb-0310-9956-ffa450edef68