Transferred from the Adobe Open Source Subversion Repository
Revision: 18765
Author:   klin@adobe.com
Date:     2010-11-21T07:18:04.332221Z

Log Message: 
-----------
Fix for RTE when sorting complex datafields. Added an internal GridSortField class that captures the full dataField when sorting on a complex dataField. This is required because sort currently RTE's when the dataField is set to a property that doesn't exist on the data object even though a sortCompareFunction is defined. I've also changed the default compare function for complex dataFields to compare the data instead of using the itemToLabel function. I've modified DataGrid to take advantage of GridSortField for figuring out when to reverse the sort of a column with a complex dataField.

QE notes: No
Doc notes: No
Bugs: SDK-28610
Reviewer: Hans
Tests run: cyclone, checkintests
Is noteworthy for integration: No

Modified Paths:
--------------
<path kind="file" action="A">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/gridClasses/GridSortField.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/gridClasses/GridColumn.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/DataGrid.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1281793 13f79535-47bb-0310-9956-ffa450edef68