Transferred from the Adobe Open Source Subversion Repository
Revision: 17445
Author:   jszeto@adobe.com
Date:     2010-08-23T18:47:40.166311Z

Log Message: 
-----------
Button icon support changes

Changed the icon support in Button. Icon is now a property instead of a style. I've added an optional skin part called iconDisplay:BitmapImage.

ButtonSkinBase does not use this skin part because BitmapImage can not be added as a child of a UIComponent. I renamed ButtonSkinBase.iconDisplay to iconInstance to avoid any confusion with the skin part name.

ButtonSkinBase listens for the "iconChange" event from the ButtonBase instead of relying on ButtonBase to push the icon to the iconDisplay.


QE notes: Need to update tests
Doc notes: none
Bugs: none
Reviewer: Corey
Tests run: checkintests
Is noteworthy for integration: yes, icon is a property now instead of a style

Modified Paths:
--------------
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/CheckBox.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/mobiletheme/src/spark/skins/mobile/supportClasses/ButtonSkinBase.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/RadioButton.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/mobiletheme/src/spark/skins/mobile/supportClasses/SelectableButtonSkinBase.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/mobiletheme/src/spark/skins/mobile/CheckBoxSkin.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/supportClasses/ButtonBase.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1281988 13f79535-47bb-0310-9956-ffa450edef68