Transferred from the Adobe Open Source Subversion Repository
Revision: 4020
Author:   rfrishbe@adobe.com
Date:     2008-11-05T00:43:18.692973Z

Log Message: 
-----------
When setItemAt() is called on a datagroup's dataProvider, we were removing the old item and adding the new one immediately.  However, when someone tries to do a swap with: getItem1, getItem2, setItemAt(item1, index2), setItemAt(item2, index1), we are temporarily in an inconsistent state.  Since this worked in Halo, to get it to work in Gumbo, we just delay any work until commitProperties().  This way, we are in a consistent state.  It could be a performance hit, but setItemAt() probably isn't used very often.

We need to go back and look at setItemAt and see what we should do with these cases where an item appears twice in the list and whether the item is a visual item or just a data item.

Also, some random ASDoc and other fixes.

QE Notes: Exclude failing tests.  File a bug for these tests.
Doc Notes: None
Bugs: SDK-16956, SDK-17909
Reviewer: Glenn

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/mx/components/DataGroup.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/mx/components/baseClasses/GroupBase.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1282088 13f79535-47bb-0310-9956-ffa450edef68