Transferred from the Adobe Open Source Subversion Repository
Revision: 4933
Author:   dsubrama@adobe.com
Date:     2009-02-12T00:15:06.489167Z

Log Message: 
-----------
A variety of fixes to Spinner and NumericStepper that exposed some inaccuracies in Range. All discovered while fixing SDK-17822. 

Additionally, added the extractValueFunction which is the sibling property to displayFormatFunction. Both can be set on a Spark NumericStepper to modify the value displayed in the text input and to extract the correct value for internal representation.

Checkintests: Pass
Mustella tests: Pass (NS, Spinner, ScrollBar, Slider)
Reviewer: Jason

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/mx/components/baseClasses/FxRange.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/mx/components/FxNumericStepper.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/mx/components/FxSpinner.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1282457 13f79535-47bb-0310-9956-ffa450edef68