Transferred from the Adobe Open Source Subversion Repository
Revision: 8959
Author:   jszeto@adobe.com
Date:     2009-07-31T17:55:21.030388Z

Log Message: 
-----------
http://bugs.adobe.com/jira/browse/SDK-22152 - Calling ensureIndexIsVisible() on Spark DropDownList is unpredictable if an item is selected

In this bug, the VScrollBar’s value starts off at 8960, is set to 220, and then set back to 8960 all before a layout pass is run.  This exposed a bug in Range where programmatically setting the value to  new value and then back to the original value resulted in the wrong value. 

The fix is to check for the value setter to compare the new value against “value” instead of “_value”. 

QE notes: Add test to Range components 
Doc notes: None
Bugs: SDK-22152
Reviewer: Hans
Tests run: NumericStepper, Slider, ScrollBar
Is noteworthy for integration: No

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/spark/src/spark/components/supportClasses/Range.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1282467 13f79535-47bb-0310-9956-ffa450edef68