Transferred from the Adobe Open Source Subversion Repository
Revision: 12709
Author:   cframpto@adobe.com
Date:     2009-12-09T13:54:43.569815Z

Log Message: 
-----------
Add support for custom context menus for RichEditableText. There are limitations to this support since it needs additional API for TextArea, TextInput, and RichEditableText to do it correctly and we need better support from TLF as well.  We should plan to do this in the next release.

For this release you can add a custom context menu to RichEditableText if it is done before the first mouse over, mouse hover or focusIn event on the RichEditableText field.  TLF will enable/disable the clipboard properties based on the editing mode and the selection.  These will show in the menu if the clipboardMenu is enabled.  

You can only remove the context menu once the component is using a flowComposer rather than a factory, which means it has had focus.

The contextMenu must be set on RichEditableText, not on TextArea or TextInput.  So for TextArea this is ta.textDisplay.contextMenu or for TextInput this is ti.textDisplay.contextMenu.  If it is set on TextArea or TextInput you won't see it since the RichEditableText menu will be displayed instead.

QE notes: 
Doc notes: see above
Bugs: partial fix for SDK-23926
Reviewer: 
Tests run: checkintests
Is noteworthy for integration:  yes - perhaps?  custom menu support for RichEditableText with severe limitations

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/spark/src/spark/components/supportClasses/RichEditableTextContainerManager.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1282512 13f79535-47bb-0310-9956-ffa450edef68