Transferred from the Adobe Open Source Subversion Repository
Revision: 13306
Author:   cframpto@adobe.com
Date:     2010-01-06T16:34:04.377095Z

Log Message: 
-----------
TLF ties activation and focus together.  If a Flex PopUp is created it is possible to get deactivate/activate events without any focus events.  If we have focus when we are activated, the selection state should be SelectionFormatState.FOCUSED not SelectionFormatState.UNFOCUSED since there might not be a follow on focusIn event.

Note that if a non-modal popop is displayed while the RET has focus, the RET is inactive but it is still possible to type into it.  The RET will not have an insertion caret until the popup is removed and the RET becomes active again.

QE notes: 
Doc notes: 
Bugs: SDK-24886
Reviewer: Alex
Tests run: checkintests, TextArea
Is noteworthy for integration:

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/spark/src/spark/components/supportClasses/RichEditableTextContainerManager.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1282515 13f79535-47bb-0310-9956-ffa450edef68