Transferred from the Adobe Open Source Subversion Repository
Revision: 17555
Author:   cframpto@adobe.com
Date:     2010-09-02T14:24:14.877787Z

Log Message: 
-----------
If RET text is editable and multiline=false, let TLF's EditManager handle the ENTER key so that if the ENTER is pressed in the middle of a list the proper list handling is done - add a new list item or close off the list, depending on position in the list - and if not in a list, the paragraph is split.

Also fix selectionHightlighting metadata since it was in the same patch.

QE notes: 
Doc notes:
Bugs: SDK-27623, SDK-25435
Reviewer: Alex
Tests run: checkintests, cyclone 
Is noteworthy for integration: no

Modified Paths:
--------------
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/RichEditableText.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/supportClasses/RichEditableTextContainerManager.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/supportClasses/SkinnableTextBase.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1282858 13f79535-47bb-0310-9956-ffa450edef68