Transferred from the Adobe Open Source Subversion Repository
Revision: 20135
Author:   cframpto@adobe.com
Date:     2011-02-04T17:17:14.255516Z

Log Message: 
-----------
Fixup ASdoc for text styles and properties in TextArea/TextInput/SkinnableTextBase which are common to spark and mobile text but have different descriptions/values.  We are restricted to how we can format the doc since there can only be one @copy and it must be last.  In addition, we do not want the summary line to be "Not supported for the mobile theme".  Stephen Gilson and I worked out a template which uses @see rather than @copy since there can be multiple @see tags.  Look at SkinnableTextBase.editable for an example when mobile supports the property and SkinnableTextBase.typicalText for an example when mobile does not support the property.


QE notes: 
Doc notes: 
Bugs: SDK-29290
Reviewer: 
Tests run: checkintests
Is noteworthy for integration:

Modified Paths:
--------------
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/framework/src/mx/core/UIComponent.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/mobilecomponents/src/spark/components/supportClasses/StyleableTextField.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/TextInput.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/styles/metadata/BasicNonInheritingTextStyles.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/styles/metadata/SelectionFormatTextStyles.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/styles/metadata/BasicInheritingTextStyles.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/styles/metadata/AdvancedInheritingTextStyles.as</path>
<path kind="file" action="M">/private/flex/sdk/branches/hero_private/frameworks/projects/spark/src/spark/components/supportClasses/SkinnableTextBase.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1282876 13f79535-47bb-0310-9956-ffa450edef68