Transferred from the Adobe Open Source Subversion Repository
Revision: 3670
Author:   gosmith@adobe.com
Date:     2008-10-15T23:26:37.608871Z

Log Message: 
-----------
The interaction of the 'text' and 'content' properties is now working as intended. If you set one and wait for the updateComplete event, you can now get the other. (In the meantime, the other is null.) If you set both at the same time, 'content' wins.

Also, setting the 'text' or the 'content' of a TextView now causes the selection range to be reset to (-1, -1), which means "no selection".

Reviewer: Deepa
Bugs: SDK-16035, SDK-16359, SDK-17307
QA: Please test the interaction of the 'text' and 'content' properties in TextGraphic, TextView, and TextArea.
Doc: No

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/mx/components/baseClasses/FxTextBase.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/mx/graphics/graphicsClasses/TextGraphicElement.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/mx/components/FxTextArea.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/mx/graphics/TextGraphic.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flex4/src/mx/components/TextView.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1282979 13f79535-47bb-0310-9956-ffa450edef68